+++ TeamCode/src/main/java/org/firstinspires/ftc/team7316/commands/AutoSlide.java
@@ -11,11 +11,11 @@ import org.firstinspires.ftc.team7316.util.copypastaLib.CombinedPath;
 public class AutoSlide extends Command {
     ElapsedTime t = new ElapsedTime();
     double lastTime;
-    int motorPower;
+    double motorPower;
     double duration;
 
 
-    public AutoSlide(int motorPower, double duration){
+    public AutoSlide(double motorPower, double duration){
         this.motorPower=motorPower;
         this.duration = duration;
     }
